[89] return 404 for request with missing grids.
Instead of returning a BadRequest and logging out no rows found, we will instead return a 404, and log nothing for missing grids.
